work on 1.21.3

+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/FuelRefineryBlockEntity.java
@@ -73,7 +73,7 @@ public class FuelRefineryBlockEntity extends BaseEnergyContainerBlockEntity impl
                 fuel = 243000 -  (int)JetSuit.Suit.getFuel(getItem(2));
             }
 
-            JetSuit.Suit.addFuel(getItem(2), (int) fuel);
+            JetSuit.Suit.addFuel(getItem(2), fuel);
             resultTank.grow(-fuel);
 
         }
